Use components-* version of modernizr package
Modernizr doesn't have a bower.json (as of version 2.8.3), and some build tools (brunch.io) depend on this to work properly (so they can see which files need to be concatenated via the "main" property). Using the components-modernizr version of the package will allow those build tools to work without any hacks/overrides.